requirements.txt
appdirs==1.4.3
click==6.7
Flask==0.12
gunicorn==19.7.1
h5py==2.7.0
itsdangerous==0.24
Jinja2==2.9.5
Keras==2.0.2
MarkupSafe==1.0
numpy==1.12.1
olefile==0.44
opencv-python==3.2.0.6
packaging==16.8
pandas==0.19.2
Pillow==4.0.0
psycopg2==2.7.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
requests==2.13.0
scikit-learn==0.18.1
scipy==0.19.0
six==1.10.0
sklearn==0.0
SQLAlchemy==1.1.7
Theano==0.9.0
Werkzeug==0.12.1
